1. file matching according to pattern 2. initial tests for workflow

+++ models/models.go
@@ -8,7 +8,7 @@ import (
 )
 
 type Model struct {
-	ID        uint `gorm:"primary_key"`
+	ID        int64 `gorm:"primary_key"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	//	DeletedAt *time.Time `sql:"index"`
